safari-600.7-branch/LayoutTests/platform/win/plugins/call-javascript-that-destroys-plugin-expected.txt

CONSOLE MESSAGE: PLUGIN: Success: executed script, and plug-in is not yet destroyed.
CONSOLE MESSAGE: PLUGIN: Plug-in destroyed.
This tests that, if a plug-in directs the browser to evaluate JavaScript that removes the plug-in from the page, the plug-in is destroyed asynchronously.
